Update Domain invariant to allow www
A domain name may be defined as www.example.com and is entirely valid.
The www is a convention and not something added by the browser etc.
Additionally, as domains can contain http such as 'http-spec.com' we
should specificially disallow http:// and https:// at the start of the
domain instead.