Reverted to scapy for sniffing
Two problems with tcpdump resulting in this shift. First, tcpdump logs
were reporting a large number of packet dropped by kernel. This seems to
have been due to the -U and --immediate-mode flags, and could be fixed
by removing these flags as well increase the buffer size and reducing
the snap length. The larger issue however was the low number of packets
captured. A quick check with wireshark indicated that the number of
packets to be captured should be in the 1000's, however tcpdump was only
managing to capture ~100. Sleeping before killing the subprocess did not
help .