Maybe suppress the OSError 9 error from scapy
For some reason the pipe used by scapy is being closed before it has a
chance to signal the background thread to close. This results in an
exception being raised. In the event that the background thread already
ran to completion after the foreground thread found the sniffer to be
running (somehow), the sniffer would no longer be in the running state
and we can safely ignore the exception caused by the race condition.